refactor maelstrom weapon tracker
